added responsiveness to the hero section
