--- src/App.css
@@ -1,3 +1,0 @@
-body{
-    background-color: #081F4A;
-}
