+++ src/CSS/herobody.css
@@ -25,3 +25,38 @@
     display: flex;
     flex-direction: row-reverse;
 }
+
+
+
+.hero_md {
+  background-color: #081F4A;
+  padding-top: 25%;
+}
+
+.hero-text_md {
+  text-align: left;
+  position: absolute;
+  top: 35%;
+  left: 45%;
+  transform: translate(-50%, -50%);
+  color: white;
+}
+
+.hero_heading_md{
+width: 280px;
+font-family: 'Lato';
+font-style: normal;
+font-weight: 700;
+font-size: 30px;
+line-height: 52px;
+
+}
+
+.hero_image_md {
+  background-image: url("../Images/Heroimage.png");
+  height: 50vh;
+  background-position: 90% 0;
+  background-repeat: no-repeat;
+  background-size:contain;
+  width: 100%;
+}
