+++ src/Component/ButtonComponent.js
@@ -10,10 +10,10 @@ const theme = createTheme({
   },
 });
 
-export default function CustomColor({variant ,spacing ,children}) {
+export default function CustomColor({variant ,spacing , size ,children}) {
   return (
     <ThemeProvider theme={theme}>
-      <Button color="neutral" variant={variant} sx={spacing}>
+      <Button color="neutral" variant={variant} sx={spacing} size={size}>
              { children }
       </Button>
     </ThemeProvider>
