+++ src/Section/Hero.js
@@ -1,13 +1,11 @@
 import React from "react";
-import HeroBody from "../Component/HeroBody";
-import Navbar from "../Component/Navbar";
+import HeroBody from "../components/herobody";
 import Header from "../components/header";
 
 function Hero() {
   return (
     <div>
       <Header />
-      {/* <Navbar /> */}
       <HeroBody />
     </div>
   );
